pyproject.toml
[project]
name = "syncthing-tray"
version = "0.1.0"
description = "Windows tray app for managing Syncthing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.0.0",
    "pyinstaller>=6.0.0",
    "pystray>=0.19.5",
]

[project.gui-scripts]
syncthing-tray = "syncthing_tray.main:main"

[project.scripts]
build-exe = "syncthing_tray.build:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/syncthing_tray"]

[tool.hatch.build.targets.wheel.force-include]
"src/syncthing_tray/syncthing.ico" = "syncthing_tray/syncthing.ico"
